--- a/.buildkite/commands/run-unit-tests.sh
+++ b/.buildkite/commands/run-unit-tests.sh
@@ -1,5 +1,7 @@
 #!/bin/bash -eu
 
+.buildkite/commands/restore-cache.sh
+
 echo "--- 🧪 Testing"
 set +e
 ./gradlew testJalapenoDebugUnitTest testDebugUnitTest