--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -41,7 +41,9 @@ steps:
           - "**/build/reports/detekt/detekt.html"
 
       - label: "lint"
-        command: .buildkite/commands/lint.sh
+        command: |
+          .buildkite/commands/restore-cache.sh
+          .buildkite/commands/lint.sh
         plugins: [$CI_TOOLKIT]
         artifact_paths:
           - "**/build/reports/lint-results*.*"
@@ -59,12 +61,14 @@ steps:
 
       - label: "🛠 Prototype Build: Mobile App"
         command: |
+          .buildkite/commands/restore-cache.sh
           ".buildkite/commands/prototype-build.sh" "WooCommerce"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
 
       - label: "🛠 Prototype Build: Wear App"
         command: |
+          .buildkite/commands/restore-cache.sh
           ".buildkite/commands/prototype-build.sh" "WooCommerce-Wear"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
@@ -74,7 +78,9 @@ steps:
     steps:
 
       - label: "Unit tests"
-        command: .buildkite/commands/run-unit-tests.sh
+        command: |
+          .buildkite/commands/restore-cache.sh
+          .buildkite/commands/run-unit-tests.sh
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :
@@ -84,7 +90,9 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "Instrumented tests"
-        command: .buildkite/commands/run-instrumented-tests.sh
+        command: |
+          .buildkite/commands/restore-cache.sh
+          .buildkite/commands/run-instrumented-tests.sh
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :