--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -41,9 +41,7 @@ steps:
           - "**/build/reports/detekt/detekt.html"
 
       - label: "lint"
-        command: |
-          .buildkite/commands/restore-cache.sh
-          .buildkite/commands/lint.sh
+        command: .buildkite/commands/lint.sh
         plugins: [$CI_TOOLKIT]
         artifact_paths:
           - "**/build/reports/lint-results*.*"
@@ -60,16 +58,12 @@ steps:
     steps:
 
       - label: "🛠 Prototype Build: Mobile App"
-        command: |
-          .buildkite/commands/restore-cache.sh
-          ".buildkite/commands/prototype-build.sh" "WooCommerce"
+        command: ".buildkite/commands/prototype-build.sh" "WooCommerce"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
 
       - label: "🛠 Prototype Build: Wear App"
-        command: |
-          .buildkite/commands/restore-cache.sh
-          ".buildkite/commands/prototype-build.sh" "WooCommerce-Wear"
+        command: ".buildkite/commands/prototype-build.sh" "WooCommerce-Wear"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
 
@@ -78,9 +72,7 @@ steps:
     steps:
 
       - label: "Unit tests"
-        command: |
-          .buildkite/commands/restore-cache.sh
-          .buildkite/commands/run-unit-tests.sh
+        command: .buildkite/commands/run-unit-tests.sh
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :
@@ -90,9 +82,7 @@ steps:
           - "**/build/test-results/merged-test-results.xml"
 
       - label: "Instrumented tests"
-        command: |
-          .buildkite/commands/restore-cache.sh
-          .buildkite/commands/run-instrumented-tests.sh
+        command: .buildkite/commands/run-instrumented-tests.sh
         plugins:
           - $CI_TOOLKIT
           - $TEST_COLLECTOR :