--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -58,12 +58,12 @@ steps:
     steps:
 
       - label: "🛠 Prototype Build: Mobile App"
-        command: ".buildkite/commands/prototype-build.sh" "WooCommerce"
+        command: ".buildkite/commands/prototype-build.sh WooCommerce"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]
 
       - label: "🛠 Prototype Build: Wear App"
-        command: ".buildkite/commands/prototype-build.sh" "WooCommerce-Wear"
+        command: ".buildkite/commands/prototype-build.sh WooCommerce-Wear"
         if: build.pull_request.id != null
         plugins: [$CI_TOOLKIT]