--- a/.buildkite/shared-pipeline-vars
+++ b/.buildkite/shared-pipeline-vars
@@ -3,5 +3,5 @@
 # This file is `source`'d before calling `buildkite-agent pipeline upload`, and can be used
 # to set up some variables that will be interpolated in the `.yml` pipeline before uploading it.
 
-export CI_TOOLKIT="automattic/a8c-ci-toolkit#3.9.0"
+export CI_TOOLKIT="automattic/a8c-ci-toolkit#9a12d43b5df24ad661ee565cdff77513dd44216c"
 export TEST_COLLECTOR="test-collector#v1.10.1"