<!-- Remember about a good descriptive title. -->

Project Thread: paaHJt-7CE-p2
Related: #13170 + #13122
Depends On: [a8c-ci-toolkit#135](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/135) -> [a8c-ci-toolkit#138](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/138)
Required By: [buildkite-ci#576](https://github.com/Automattic/buildkite-ci/pull/576)

<!-- Id number of the GitHub issue this PR addresses. -->

---

### Description
<!-- Take the time to write a good summary. Why is it needed? What does it do? When fixing bugs try to avoid just writing “See original issue” – clarify what the problem was and how you’ve fixed it. -->

This PR enables dependency cache on CI for this project. 💾

The dependency cache mechanism is split between:
- A weekly scheduled **save cache** build (see [commit](https://github.com/woocommerce/woocommerce-android/pull/13303/commits/684c6e6d8d1a8b3942705a10e4a43c3b2f7a484c) and [buildkite-ci#576](https://github.com/Automattic/buildkite-ci/pull/576) PR), which targets the below tasks only:
  - Mobile App (`WooCommerce:assembleJalapenoDebug`)
  - Wear App (`WooCommerce-Wear:assembleJalapenoDebug`)
  - Lint (`WooCommerce:lintJalapenoDebug`)
  - Unit Tests (`ssembleJalapenoDebugUnitTest assembleDebugUnitTest`)
  - Android Tests (`assembleJalapenoDebugAndroidTest`)
- A **restore cache** step that is run on targeted [pipeline.yml](https://github.com/woocommerce/woocommerce-android/blob/trunk/.buildkite/pipeline.yml) related jobs, via their corresponding individual scripts:
  - Mobile App (`prototype-build.sh`)
  - Wear App (`prototype-build.sh`)
  - Lint (`lint.sh`)
  - Unit Tests (`run-unit-tests.sh`)
  - Android tests (`run-instrumented-tests.sh`)

PS.1: The `detekt` job isn't targeted because it only takes about 20+ seconds of network request to download all the dependencies needed for this specific job ([build](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e466-4438-be6a-0a67488f8395) + [scan](https://scans.gradle.com/s/ftu4zklrffa3q/performance/network-activity)), which is about the same time it take for the dependency cache to actually be restored, which is more or less about 10+ seconds ([build](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c246-4c75-9808-31925b53689b/189-198)). Thus, this optimization is not really worth it for this specific job. For all those other jobs targeted, it takes more than a minute of network requests to download all the dependencies needed ([build](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e46c-4e6b-9157-596bd033d26f) + [scan](https://scans.gradle.com/s/fbw7fyc4enqw2/performance/network-activity)), and as such the diff is worth it.

PS.2: Dependency cache isn't enabled on the [release-builds.yml](https://github.com/woocommerce/woocommerce-android/blob/trunk/.buildkite/release-builds.yml) pipeline just yet, but just temporarily, and until we become comfortable with running this new process on the main [pipeline.yml](https://github.com/woocommerce/woocommerce-android/blob/trunk/.buildkite/pipeline.yml) pipeline.

---

### Future Improvement

As per [this](https://github.com/woocommerce/woocommerce-android/pull/13303#discussion_r1916744733) comment, me and @wzieba has been discussing on DMs about some possible future improvement, which are including but are not limited to the below list:

1. Only execute the `assemble android test` task, which depends on `assemble` task anyway, and would most likely resolve most of the dependencies, for all other tasks to benefit from. However, this needs further testing to make sure that running other tasks such us `lint` and `assemble unit test` are providing little to no help.
2. Create a custom Gradle task and utilize Gradle's [init scripts](https://docs.gradle.org/current/userguide/init_scripts.html) (`init.d`). Using the init script, go through all available modules for a project, and then, using that custom Gradle task, resolve all dependencies for all tasks that are available per module. We could take some inspiration from plugins such as [this](https://plugins.gradle.org/plugin/org.gradle.github-dependency-graph-gradle-plugin).
3. Instead of scheduling the `dependency cache` build to run weekly ([buildkite-ci#576](https://github.com/Automattic/buildkite-ci/pull/576)), make it run more often, possibly scheduling it to even run daily. However, since dependencies are not being updated too often, and, even when they do on that specific week, it would only be for a few of them. This will indeed cause some extra network requests, but most of the dependencies cached would still benefit from the existing and weekly refreshed dependency cache.
4. Enable dependency cache on the [release-builds.yml](https://github.com/woocommerce/woocommerce-android/blob/trunk/.buildkite/release-builds.yml) pipeline.

---

### Testing information
<!-- This is your opportunity to break out individual scenarios that need testing (when necessary) and/or include a checklist for the reviewer to go through. Consider documenting the following from your own completed testing: devices used, alternate workflows, edge cases, affected areas, critical flows, areas not tested, and any remaining unknowns. Provide feedback on this new section of the PR template through Sept 30, 2024 to Apps Quality; additional context here: p91TBi-b8z-p2#comment-12036 -->

This dependency cache solution was thoroughly tested via this #13170 test-only PR, but to help you asserting that this change is indeed making an impact, plus not causing any trouble for the project, try and follow the below:

- `Save Cache`:

  - Triggering [schedules/dependency-cache.yml](https://github.com/woocommerce/woocommerce-android/pull/13303/files#diff-f1eb78f1f1134bf64e6d4a27fd0b816ae7dff3a0da45f4115e8898bc6939642e) manually via [this](https://buildkite.com/automattic/woocommerce-android/builds/26606) build:
    - Starts [downloading the dependencies per task](https://buildkite.com/automattic/woocommerce-android/builds/26606#019469da-aebc-4dc3-a4d8-e3b967c76b41/187).
    - Caches [the dependencies successfully](https://buildkite.com/automattic/woocommerce-android/builds/26606#019469da-aebc-4dc3-a4d8-e3b967c76b41/9110-9111).
      - Duration: [42s](https://buildkite.com/automattic/woocommerce-android/builds/26606#019469da-aebc-4dc3-a4d8-e3b967c76b41/9110-9121)
      - Size: [625.74 MB](https://buildkite.com/automattic/woocommerce-android/builds/26606#019469da-aebc-4dc3-a4d8-e3b967c76b41/9110-9122)
    - Total time it takes for the `dependency cache` job to complete: [7m 42s](https://buildkite.com/automattic/woocommerce-android/builds/26606#019469da-aebc-4dc3-a4d8-e3b967c76b41)
  - Sending a Slack notification (`#android-core-notifs`):
    - On success, notifying that: `The dependency cache has been updated.` (p1736943241087289-slack-C0787KRDYDC)
    - On failure, notifying that: `Failure to update the dependency cache.` (not tested, but should behave similarly to ☝️)
- `Restore Cache`:

| Job | BT (Without) | BT (With) | BT (Diff*) | NR (Without) | NR (With) | RC (Without) | RC (With) | Win |
| ---- | ------------ | ---------- | ---- | ------------- | --------- | -------------- | --------- | ---- |
| Mobile App | [5m 38s](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e46c-4e6b-9157-596bd033d26f) | [3m 53s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c246-4c75-9808-31925b53689b) | `- 1m 45s` | [1m 15s](https://scans.gradle.com/s/fbw7fyc4enqw2/performance/network-activity) | [0s](https://scans.gradle.com/s/xltgddi566p4g/performance/network-activity) | `0s` | [12s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c246-4c75-9808-31925b53689b/189-198) | `1m 3s` |
| Wear App | [4m 43s](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e46d-4415-8e6f-0a2f6245333b) | [2m 56s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c247-4869-b846-4a69ac9675c1) | `- 1m 47s` | [59s](https://scans.gradle.com/s/6hpqcu2spugwg/performance/network-activity) | [0s](https://scans.gradle.com/s/tfyjvmh33szxe/performance/network-activity) | `0s` | [15s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c247-4869-b846-4a69ac9675c1/189-198) | `44s` |
| Lint | [6m 9s](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e469-4f0b-8c79-0bbf0aef92c1) | [4m 6s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c242-4629-8651-6b240d97f4de) | `- 2m 3s` | [1m 49s](https://scans.gradle.com/s/3vodnbxajkqug/performance/network-activity) | [0s](https://scans.gradle.com/s/nstv7ejuxme4y/performance/network-activity) | `0s` | [14s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c242-4629-8651-6b240d97f4de/189-198) | `1m 35s` |
| Unit Test | [6m 0s](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e470-4846-b3bf-3c65b0f3e6c3) | [3m 57s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c24a-4bb8-9a3e-24052f20e33c) | `- 2m 3s` | [1m 31s](https://scans.gradle.com/s/uusmmn4gmwymw/performance/network-activity) | [0s](https://scans.gradle.com/s/5u33enw43hjec/performance/network-activity) (ish) | `0s` | [10s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c24a-4bb8-9a3e-24052f20e33c/224-233) | `1m 21s` |
| UI Test | [10m 3s](https://buildkite.com/automattic/woocommerce-android/builds/26561#01946578-e471-4783-abf1-c5f08f04612f) | [13m 37s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c24a-454f-a98f-f81c82f2c908) | `+ 3m 34s` | [1m 38s](https://scans.gradle.com/s/stvpooxfp3bnm/performance/network-activity) | [0s](https://scans.gradle.com/s/eaedij3zqpbio/performance/network-activity) | `0s` | [15s](https://buildkite.com/automattic/woocommerce-android/builds/26609#019469e4-c24a-454f-a98f-f81c82f2c908/224-233) | `1m 23s` |

> [!NOTE]
> - `Without`: Without Dependency Cache
> - `With`: With Dependency Cache
> - `BT`: Build Time
> - `NR`: Network Requests
> - `RC`: Restore Cache
> - `Win`: NR (Without) - RC (With)

> [!WARNING]
> (*): The `BT (Diff)` presented above is relative, so don't read too much into it, it might not match with the `WIN`, which is a much more solid number. But still, it is nice to have and compare those two number together. To to be clear about that, although I tried to pick the most representative builds, without and with the dependency cache mechanism in place, there are other factors that can make any build faster, or slower for that matter (like `build cache`, `firebase test lab`, etc), so again, don't read too much into it.
 
---

### Merge Instructions

- [x] Wait for [a8c-ci-toolkit#135](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/135) -> [a8c-ci-toolkit#138](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/138) to be reviewed, approved and merged.
- [x] Create a new [a8c-ci-toolkit release](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/releases) ([3.9.0](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/releases/tag/3.9.0) -> [3.9.1](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/releases/tag/3.9.1) ).
- [x] Update [CI_TOOLKIT](https://github.com/woocommerce/woocommerce-android/blob/trunk/.buildkite/shared-pipeline-vars#L6) to point to this newly created [a8c-ci-toolkit release](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/releases) (3547e664e43bcb37d7e5c67190aa68bf7f492842 -> 6ac4eb4555927d660994fb8bb7e2fc3aae6faad6).
- [x] Remove `status: do not merge` label.
- [ ] Merge this PR.
- [ ] Close #13122 and #13170 test-only PRs.

---

- [x] I have considered if this change warrants release notes and have added them to `RELEASE-NOTES.txt` if necessary. Use the "[Internal]" label for non-user-facing changes.

## Reviewer (or Author, in the case of optional code reviews):

Please make sure these conditions are met before approving the PR, or request changes if the PR needs improvement:

- [x] The PR is small and has a clear, single focus, or a valid explanation is provided in the description. If needed, please request to split it into smaller PRs.
- [x] Ensure Adequate Unit Test Coverage: The changes are reasonably covered by unit tests or an explanation is provided in the PR description.
- [x] Manual Testing: The author listed all the tests they ran, including smoke tests when needed (e.g., for refactorings). The reviewer confirmed that the PR works as expected on big (tablet) and small (phone) in case of UI changes, and no regressions are added.

<!-- Pull request guidelines: https://github.com/woocommerce/woocommerce-android/blob/develop/docs/pull-request-guidelines.md -->